+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add preliminary support for automatic fixing of violations, starting with `require-default-empty-arrays`.
+
 ## 0.3.0 (2024-04-03)
 
 - [#330](https://github.com/loop-payments/prisma-lint/issues/330) Remove `requirePrefix` option and add `requireUnderscorePrefixForIds` to `field-name-mapping-snake-case` to actually support MongoDB naming conventions.